Refactoring, adjusting padding
