+++ packages/lib/components/shared/side-menu-shared.test.ts
@@ -1,8 +1,7 @@
-import { FolderEntity } from '../../services/database/types';
 import { getTrashFolder, getTrashFolderId } from '../../services/trash';
-import { buildFolderTree, renderFolders } from './side-menu-shared';
+import { buildFolderTree, RenderFolderItemEvent, renderFolders } from './side-menu-shared';
 
-const renderItem = (folder: FolderEntity, hasChildren: boolean, depth: number) => {
+const renderItem = ({ folder, hasChildren, depth }: RenderFolderItemEvent) => {
 	return [folder.id, hasChildren, depth];
 };
